Updated imports, refactored replace function call inside task-card.js
